interview_coach/requirements.txt
streamlit==1.28.0
gradio==3.50.0
python-dotenv==1.0.0
camel-ai==0.1.0
openai==1.1.0
markdown==3.4.3
docx2markdown==0.0.3
pandas==2.1.0
numpy==1.24.3
typer==0.9.0
rich==13.4.2